feat: split cmd/main.go into focused per-concern files
Split the 455-LOC cmd/main.go into five files under the 200-LOC budget:

- cmd/main.go (120 LOC) — main, init, before, handleSignals, runtime client
  package vars, version vars, constants
- cmd/runtime.go (109 LOC) — createRuntimeClient, tlsConfig, runtime factory vars
- cmd/logging.go (40 LOC) — setupLogging helper consumes cliflags.Flags
- cmd/flags.go (102 LOC) — globalFlags
- cmd/commands.go (122 LOC) — initializeCLICommands (extracted from flags.go to
  stay under budget; combined would have been 217 LOC)

before() now calls setupLogging(cliflags.NewV1FromApp(c)) and delegates to
createRuntimeClient. No public surface changes; same package main, same call
graph. All existing tests pass; lint clean.
